Move logging helpers to dedicated class
